requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
psycopg2
Click
flake8
python-dotenv
flask-debugtoolbar
